aws/lambda/cross_repo_ci_relay/result/requirements.txt
requests
PyYAML
PyGithub
PyJWT[crypto]
clickhouse-connect
python-dotenv
redis
